Removed redundant casting to int

+++ src/pl/jasonxiii/pong/Methods.java
@@ -4,6 +4,6 @@ public final class Methods
 {
 	public static int clampInt(int a, int n, int b)
 	{
-		return (int)Math.min(Math.max(n, a), b);
+		return Math.min(Math.max(n, a), b);
 	}
 }
